cleanup
- add/use 'debug' module
- refactor some of the code out of the big api.js file
