+++ lib/api.js
@@ -8,55 +8,17 @@ var _ = require('lodash');
 var assert = require('assert');
 var qs = require('querystring');
 var stringify = require('json-stable-stringify');
+var debug = require('debug')('gp:api');
 
 // protobuf initialization
 var builder = ProtoBuf.loadProtoFile(__dirname + '/data/googleplay.proto');
 var ResponseWrapper = builder.build("ResponseWrapper");
 var PreFetch = builder.build("PreFetch");
 var BulkDetailsRequest = builder.build("BulkDetailsRequest");
-
-function getOrElseThrow(thing, msg) {
-  if (typeof thing === 'undefined') {
-    throw new Error(msg);
-  }
-}
-_.mixin({'getOrElseThrow': getOrElseThrow});
-
-function LoginError(msg) {
-  Error.call(this);
-  Error.captureStackTrace(this, arguments.callee);
-  this.message = msg;
-  this.name = 'LoginError';
-}
-
-LoginError.prototype = Object.create(Error.prototype);
-LoginError.prototype.constructor = LoginError;
-
-function RequestError(msg) {
-  Error.call(this);
-  Error.captureStackTrace(this, arguments.callee);
-  this.message = msg;
-  this.name = 'RequestError';
-}
-
-RequestError.prototype = Object.create(Error.prototype);
-RequestError.prototype.constructor = RequestError;
-
-/**
- * Parsed string into object.
- * @param {String} lines - e.g., "FOO=bar\nBAZ=foo"
- * @return {Object} parsed object result, e.g., {foo: "bar", baz: "foo"}
- */
-function responseToObj(lines) {
-  return _.chain(lines.split('\n')).reduce(function (obj, line) {
-    var pair = line.split('=');
-    assert(pair.length == 2, 'expected list of pairs from server');
-    var key = pair[0].toLowerCase();
-    var val = pair[1];
-    obj[key] = val;
-    return obj;
-  }, {}).value();
-}
+var getOrElseThrow = require('./get-or-throw');
+var responseToObj = require('./response-to-obj');
+var RequestError = require('./errors').RequestError;
+var LoginError  = require('./errors').LoginError;
 
 
 /**
@@ -67,7 +29,7 @@ function responseToObj(lines) {
  * @param {String} password - required
  * @param {String} androidId - required
  * @param {Boolean} useCache - enable debug output (default: true)
- * @param {Boolean} debug - enable debug output (default: false)
+ * @param {Boolean} debug - enable request debug output (default: false)
  * @return {type}
  */
 var GooglePlay = (function GooglePlay(username, password, androidId, useCache, debug) {
@@ -82,9 +44,9 @@ var GooglePlay = (function GooglePlay(username, password, androidId, useCache, d
     require('request-debug')(request);
   }
 
-  _.getOrElseThrow(username, 'Require username');
-  _.getOrElseThrow(password, 'Require password');
-  _.getOrElseThrow(androidId, 'Require Android ID');
+  getOrElseThrow(username, 'Require username');
+  getOrElseThrow(password, 'Require password');
+  getOrElseThrow(androidId, 'Require Android ID');
 
   var DEVICE_COUNTRY, OPERATOR_COUNTRY, LOGIN_LANGUAGE;
   DEVICE_COUNTRY = LOGIN_LANGUAGE = OPERATOR_COUNTRY = "us";
@@ -270,15 +232,14 @@ var GooglePlay = (function GooglePlay(username, password, androidId, useCache, d
         memoizedExecuteRequestApi2.cache[cacheKey] = Promise.resolve(entry.response);
         if (ttl) {
           setTimeout(function () {
-            if (debug) {
-              console.log('invalidating cache key: %s', cacheKey);
-            }
+            debug('invalidating cache key: %s', cacheKey);
             delete memoizedExecuteRequestApi2.cache[cacheKey];
           }, ttl).unref();
         }
       }
     });
   }
+
   /**
    * Convert a data buffer to a ResponseWrapper object.
    * @param {Buffer} data
@@ -428,17 +389,13 @@ var GooglePlay = (function GooglePlay(username, password, androidId, useCache, d
     if (!memoizedExecuteRequestApi2.cache) {
       return;
     }
-    if (debug) {
-      console.log('invalidating cache');
-      console.log('old keys: %s', cachedKeys());
-    }
+    debug('invalidating cache');
+    debug('old keys: %s', cachedKeys());
     memoizedExecuteRequestApi2.cache.each(function (v, k) {
       delete memoizedExecuteRequestApi2.cache[k];
     });
 
-    if (debug) {
-      console.log('now keys: %s', cachedKeys());
-    }
+    debug('now keys: %s', cachedKeys());
   }
 
   return {
