+++ lib/get-or-throw.js
@@ -1,0 +1,6 @@
+module.exports = function getOrElseThrow(thing, msg) {
+  if (typeof thing === 'undefined') {
+    throw new Error(msg);
+  }
+};
+
