+++ package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "bluebird": "~2.3.6",
+    "debug": "^2.2.0",
     "json-stable-stringify": "~1.0.0",
     "lodash": "~2.4.1",
     "protobufjs": "git+https://github.com/dweinstein/ProtoBuf.js.git#int64-to-number",
